+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Fix `comp_precache_sound` not handling sound characters at the start of raw filenames.
 * Allow `comp_vactube_start` to be set to have a timer which starts disabled.
 * If required, add various QC flags like `$mostlyopaque` to propcombined props.
+* `comp_pack_rename` will now automatically also include the auxiliary `.mdl` files.
 * Limit the size of propcombined groups to avoid hitting vertex limits.
 * Weapon scripts are now packed along with the models/sounds they use.
 * Prevent automatically packing models specified only as Hammer previews in various entities.